requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
aws-cdk-lib==2.213.0
aws-cdk.asset-awscli-v1==2.2.242
aws-cdk.asset-kubectl-v20==2.1.4
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.aws-lambda-python-alpha==2.213.0a0
aws-cdk.cloud-assembly-schema==48.6.0
aws-requests-auth==0.4.3
beautifulsoup4==4.13.5
boto3==1.40.21
botocore==1.40.21
cattrs==25.1.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
constructs==10.3.0
dill==0.4.0
dnspython==2.7.0
docstring_parser==0.17.0
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
frozenlist==1.7.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jmespath==1.0.1
jsii==1.113.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
lxml==6.0.1
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.2
mcp==1.13.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
opentelemetry-api==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-threading==0.57b0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
publication==0.0.3
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-stacks==0.0.1
pywin32==311
PyYAML==6.0.2
readabilipy==0.3.0
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rpds-py==0.27.1
s3transfer==0.13.1
sentry-sdk==2.35.1
shellingham==1.5.4
six==1.17.0
slack_bolt==1.24.0
slack_sdk==3.36.0
sniffio==1.3.1
soupsieve==2.8
sse-starlette==3.0.2
starlette==0.47.3
strands-agents==1.6.0
strands-agents-tools==0.2.5
sympy==1.14.0
tenacity==9.1.2
typeguard==2.13.3
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
